make asset modifcation to be done only on prod env
