+++ server/src/app.module.ts
@@ -39,8 +39,6 @@ import { GroupPermissionsModule } from './modules/group_permissions/group_permis
 import { TooljetDbModule } from './modules/tooljet_db/tooljet_db.module';
 import { PluginsModule } from './modules/plugins/plugins.module';
 import { CopilotModule } from './modules/copilot/copilot.module';
-import * as path from 'path';
-import * as fs from 'fs';
 import { AppEnvironmentsModule } from './modules/app_environments/app_environments.module';
 import { RequestContextModule } from './modules/request_context/request-context.module';
 import { ScheduleModule } from '@nestjs/schedule';
@@ -101,27 +99,7 @@ const imports = [
   CopilotModule,
 ];
 
-if (process.env.SERVE_CLIENT !== 'false') {
-  const filesToReplaceAssetPath = ['index.html', 'runtime.js', 'main.js'];
-
-  for (const fileName of filesToReplaceAssetPath) {
-    const file = join(__dirname, '../../../', 'frontend/build', fileName);
-
-    let newValue = process.env.SUB_PATH;
-
-    if (process.env.SUB_PATH === undefined) {
-      newValue = fileName === 'index.html' ? '/' : '';
-    }
-
-    const data = fs.readFileSync(file, { encoding: 'utf8' });
-
-    const result = data
-      .replace(/__REPLACE_SUB_PATH__\/api/g, path.join(newValue, '/api'))
-      .replace(/__REPLACE_SUB_PATH__/g, newValue);
-
-    fs.writeFileSync(file, result, { encoding: 'utf8' });
-  }
-
+if (process.env.SERVE_CLIENT !== 'false' && process.env.NODE_ENV === 'production') {
   imports.unshift(
     ServeStaticModule.forRoot({
       // Have to remove trailing slash of SUB_PATH.
